opus/opus1_amplify/anthology/repositories/anthology-ai-publishing/requirements.txt
google-cloud-aiplatform>=1.25.0
google-cloud-scheduler>=2.7.3
google-cloud-secret-manager>=2.12.6
google-cloud-storage>=2.7.0
google-cloud-monitoring>=2.14.1
PyYAML>=6.0
anthropic>=0.3.0
openai>=0.27.8
pytest>=7.3.1
requests>=2.28.2